+++ src/library/systemObjects/include/sys.h
@@ -11,8 +11,8 @@
 
 class sys
 {
-    const int                          numOfMolTypes;
-    std::vector<int>                   numOfMol;
+    const size_t                       numOfMolTypes;
+    std::vector<size_t>                numOfMol;
     std::vector<std::vector<molecule>> molecules;
     double                             sysEnergy;
 
@@ -24,17 +24,17 @@ class sys
     sys operator=(const sys&) = delete;
     sys operator=(sys&&) = delete;
 
-    sys(const int, std::vector<int>, const std::vector<molecule>&);
+    sys(const size_t, std::vector<size_t>, const std::vector<molecule>&);
 
-    std::vector<int>&                   getNumOfMol() { return numOfMol; };
+    std::vector<size_t>&                getNumOfMol() { return numOfMol; };
     std::vector<std::vector<molecule>>& getMols() { return molecules; };
     void                                resetEnergy() { sysEnergy = 0; }
     void addEnergy(double inEn) { sysEnergy += inEn; }
     void resetAllForces();
 
-    int getNumOfMolTypes() const { return numOfMolTypes; }
+    size_t getNumOfMolTypes() const { return numOfMolTypes; }
 
-    const std::vector<int>& getNumOfMol() const { return numOfMol; };
+    const std::vector<size_t>& getNumOfMol() const { return numOfMol; };
     const std::vector<std::vector<molecule>>& getMols() const
     {
         return molecules;
